requirements.txt
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
Django==3.0.8
djangorestframework==3.11.0
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==20.0.4
idna==2.10
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
numpy==1.19.0
opencv-python==4.2.0.34
parso==0.7.0
pickleshare==0.7.5
pillow==7.2.0
prompt-toolkit==3.0.5
pyfcm==1.4.7
Pygments==2.6.1
PyMySQL==0.9.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
